svg/painting/parsing/color-rendering-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing color-rendering with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#ColorRenderingProperty" rel="help"/><h:meta content="color-rendering supports the full grammar 'auto | optimizeSpeed | optimizeQuality'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("color-rendering", "auto");
test_valid_value("color-rendering", "optimizespeed");
test_valid_value("color-rendering", "optimizequality");

  ]]></script></svg>
